Correct index of linear interp
The results cannot be perfect for both interpolations at the same time, since they depend
on the pixels you take as origin. For linear interp vectorized they took 0,0 while
for bilinear they actually took 1,1. I wanted to make bilinear the closer to the solution
,since it is the one I will use in the end, therefore I will use as origin 1,1.
Hopefully, this serves as a proof that my decision is rational, in case the TAs decide
that my sol. is wrong...